requirements.txt
langchain 
langchainhub
langchain_unstructured
langchain-community
langchain-huggingface
langchain-openai
langchain-ollama
langchain_chroma
chromadb 
transformers 
sentence-transformers
ipykernel
python-dotenv
pypdf
openai 
unstructured
libmagic
loguru